Move FeedbackForm to shared components and fix console error by using React.Fragment instead of array
