+++ services/QuillLMS/client/app/bundles/Grammar/components/conceptsFeedback/conceptFeedback.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { connect } from 'react-redux'
-import { ConceptExplanation } from '../../../Shared/index'
+import { ConceptExplanation, FeedbackForm } from '../../../Shared/index'
 import { ActionTypes } from '../../actions/actionTypes'
 import * as actions from '../../actions/conceptsFeedback'
 import { Concept } from '../../interfaces/concepts'
@@ -8,7 +8,6 @@ import { ConceptFeedback } from '../../interfaces/conceptsFeedback'
 import { Match } from '../../interfaces/match'
 import { ConceptsFeedbackState } from '../../reducers/conceptsFeedbackReducer'
 import { ConceptReducerState } from '../../reducers/conceptsReducer'
-import FeedbackForm from './feedbackForm'
 
 interface ConceptFeedbackComponentProps {
   dispatch: Function;
